+++ src/Exceptions/DifferentModulusError.php
@@ -2,20 +2,20 @@
 namespace Marcoconsiglio\ModularArithmetic\Exceptions;
 
 use ErrorException;
-use Marcoconsiglio\ModularArithmetic\ModularInteger;
+use Marcoconsiglio\ModularArithmetic\ModularNumber;
 
 /**
- * The exception thrown when two number have different modules.
+ * The exception thrown when two number have different modulus.
  */
 class DifferentModulusError extends ErrorException
 {
     /**
      * Construct the exception with the two number.
      */
-    public function __construct(ModularInteger $a, ModularInteger $b)
+    public function __construct(ModularNumber $a, ModularNumber $b)
     {
         return parent::__construct(
-            "Different modules cannot be used ($a->modulus and $b->modulus)."
+            "Two different modulus cannot be used ($a->modulus and $b->modulus)."
         );
     }
 }