+++ tests/Unit/Exceptions/DifferentModulusErrorTest.php
@@ -5,25 +5,25 @@ use PHPUnit\Framework\Attributes\CoversClass;
 use PHPUnit\Framework\Attributes\TestDox;
 use PHPUnit\Framework\Attributes\UsesClass;
 use Marcoconsiglio\ModularArithmetic\Exceptions\DifferentModulusError;
-use Marcoconsiglio\ModularArithmetic\ModularInteger;
+use Marcoconsiglio\ModularArithmetic\ModularNumber;
 use Marcoconsiglio\ModularArithmetic\Tests\Unit\TestCase;
 
 #[CoversClass(DifferentModulusError::class)]
-#[UsesClass(ModularInteger::class)]
+#[UsesClass(ModularNumber::class)]
 #[TestDox("The DifferentModulusError")]
 class DifferentModulusErrorTest extends TestCase
 {
-    #[TestDox("is thrown when two ModularInteger have different modulus.")]
+    #[TestDox("is thrown when two ModularNumber have different modulus.")]
     public function test_different_modulus_error(): void
     {
         // Arrange
         $a_modulus = 10;
         $b_modulus = 20;
-        $a = new ModularInteger(
+        $a = new ModularNumber(
             $this->randomInteger(),
             $a_modulus
         );
-        $b = new ModularInteger(
+        $b = new ModularNumber(
             $this->randomInteger(),
             $b_modulus
         );
@@ -33,7 +33,7 @@ class DifferentModulusErrorTest extends TestCase
 
         // Assert
         $this->assertEquals(
-            "Different modules cannot be used ($a->modulus and $b->modulus).",
+            "Two different modulus cannot be used ($a->modulus and $b->modulus).",
             $error->getMessage()
         );
     }